inside Scrollpane class : int userNum = theBot.getUnP().getTotalUserCount(); Object[][] data = new Object[userNum][5];
Problem was, getTotalUserCount() was used to find total usercount for rank. but when being called, inside the MyBotUserPoints class,  the usermap was being updated with the updateUserMap() method, the rank was not being Initialized/Updated, so if it was null, and return = -1. So i made sure that the updateRankedList() method is being called in the updateUserMap() method, and all seems well.